Added Shop functionality
Version 1 of the shop window, fully works and supports both buying and selling items as well as item stacks, and rendering the players inventory in the shop window. Still WIP waiting on checks for if the players inventory is full so the items can be added directly but that functionally is planned for a future update to the inventory system that addresses dropping items.