House Building First Draft
The Basics of my house building system, Still needs a bit of tidying up and testing to make sure it performs properly in edge cases but so far its stable and working as intended, it works by enabling TileMaps that correspond  to a room. I have it set up so that the GUI is activated by a workbench outside the house but you could modify it and add a cutscene if you wanted, because if the player or another Rigidbody is inside the wall they will currently get stuck, but because that doesnt apply to my use case I have not included a fix.