Its time to SAVE THE WORLD!!!
Implemented saving of world data as well as player data. Also implemented Load System WIP currently working on both player stats and persistent objects tracked by the Persistence Controller script.

Next is to work on saving the players inventory and equipment, which will most likely require some changes to how the game tracks everything as the inventory system is using scriptable objects at the moment and things like the item amount are static to the instance of the item type.